+++ care-point-account/CarePointAccount/src/site/app/master/item/item-controller.js
@@ -67,15 +67,22 @@
 
                 //save item
                 $scope.ui.saveItems = function () {
+                    var check = true;
                     if (!$scope.model.itemData.type) {
+                        check = false;
                         Notification.error("enter item type");
-                    } else if (!$scope.model.itemData.name) {
+                    }
+                    if (!$scope.model.itemData.name) {
+                        check = false;
                         Notification.error("enter item name");
-                    } else if ($scope.model.itemData.type
-                            && $scope.model.itemData.name) {
-                        ConfirmPane.primaryConfirm("Save Item !")
+                    }
+                    if (!$scope.model.userPermission.add) {
+                        ch = false;
+                        Notification.error('you have no permission to save new item !');
+                    }
+                    if (check) {
+                        ConfirmPane.primaryConfirm("do you want to save Item !")
                                 .confirm(function () {
-
                                     $scope.model.saveItem()
                                             .then(function () {
                                                 Notification.success("Item save Success");
@@ -90,113 +97,58 @@
                     }
                 };
 
-                $scope.ui.savePriceCategoryDetail = function () {
-                    if (!$scope.model.priceCategoryDetail.item) {
-                        Notification.error("enter item type");
-
-                    } else if (!$scope.model.priceCategoryDetail.priceCategory) {
-                        Notification.error("enter item name");
-
-                    } else if (!$scope.model.priceCategoryDetail.normalPrice) {
-                        Notification.error("enter item sale price normal");
-
-                    } else if (!$scope.model.priceCategoryDetail.registerPrice) {
-                        Notification.error("enter item sale price register");
-
-                    } else if ($scope.model.priceCategoryDetail.item
-                            && $scope.model.priceCategoryDetail.priceCategory
-                            && $scope.model.priceCategoryDetail.normalPrice
-                            && $scope.model.priceCategoryDetail.registerPrice) {
-                        var requestData = $scope.model.duplicateCheckPriceCategoryDetails($scope.model.priceCategoryDetail.item, $scope.model.priceCategoryDetail.priceCategory);
-                        if (angular.isUndefined(requestData))
-                            ConfirmPane.primaryConfirm("Save price categiry details !")
-                                    .confirm(function () {
-                                        $scope.model.savePriceCategoryDetail()
-                                                .then(function () {
-                                                    Notification.success("save price categiry details");
-                                                    $scope.model.priceCategoryDetail.item = $scope.selectNextPriceCategoryDetails;
-                                                }, function () {
-                                                    Notification.error("save price categiry details fail");
-                                                });
-                                    })
-                                    .discard(function () {
-                                        console.log('discard');
-                                    });
-
-                    } else {
-                        Notification.error("This Item Price Category Details Allrady Exists !");
-                    }
-                };
-
-                $scope.ui.editePriceCategoryDetail = function (priceCategoryDetail, $index) {
-                    $scope.model.editePriceCategoryDetail(priceCategoryDetail, $index);
-                };
-
-                $scope.ui.selectNextPriceCategoryDetails = function (value) {
-                    if (value) {
-                        $scope.selectNextPriceCategoryDetails = $scope.model.priceCategoryDetail.item;
-                    } else {
-                        $scope.selectNextPriceCategoryDetails = null;
-                    }
-                };
-
                 //save consumable item
                 $scope.ui.saveConsumable = function () {
-                    ConfirmPane.primaryConfirm("Save this Consumable Item !")
-                            .confirm(function () {
-                                $scope.model.saveConsumable()
-                                        .then(function () {
-                                            Notification.success("Consumable Item Save Success");
-                                        }, function () {
-                                            Notification.error("Consumable Item Save Fail");
-                                        });
+                    var check = true;
+                    if (!$scope.model.userPermission.add) {
+                        check = false;
+                        Notification.error('you have no permission to save new Consumable Item !');
+                    }
+                    if (check) {
+                        ConfirmPane.primaryConfirm("Save this Consumable Item !")
+                                .confirm(function () {
+                                    $scope.model.saveConsumable()
+                                            .then(function () {
+                                                Notification.success("Consumable Item Save Success");
+                                            }, function () {
+                                                Notification.error("Consumable Item Save Fail");
+                                            });
 
-                            })
-                            .discard(function () {
-                                console.log('discard');
-                            });
+                                })
+                                .discard(function () {
+                                    console.log('discard');
+                                });
+                    }
                 };
 
-                //save item check detail
-                $scope.ui.saveItemCheckDetail = function () {
-                    ConfirmPane.primaryConfirm("Save this Detail !")
-                            .confirm(function () {
-                                $scope.model.saveItemChechDetail()
-                                        .then(function () {
-                                            Notification.success("Item Check detail Save Success");
-                                        }, function () {
-                                            Notification.error("Item Check detail Save Fail");
-                                        });
 
-                            })
-                            .discard(function () {
-                                console.log('discard');
-                            });
-
-                };
 
                 //edit item
                 $scope.ui.editeItems = function (items, $index) {
-                    $scope.textViewMode = items.type;
-                    $scope.model.editeItem(items, $index);
+                    var check = true;
+                    if (!$scope.model.userPermission.update) {
+                        ch = false;
+                        Notification.error('you have no permission to edit Item !');
+                    }
+                    if (check) {
+                        $scope.textViewMode = items.type;
+                        $scope.model.editeItem(items, $index);
+                    }
                 };
 
                 //delete item
                 $scope.ui.deleteItems = function (items, $index) {
-                    $scope.model.deleteItem(items, $index);
+                    var check = true;
+                    if (!$scope.model.userPermission.delete) {
+                        ch = false;
+                        Notification.error('you have no permission to delete Item !');
+                    }
+                    if (check) {
+                        $scope.model.deleteItem(items, $index);
+                    }
                 };
 
-                //delete item
-                $scope.ui.deletePriceCategoryDetail = function (items, $index) {
-                    ConfirmPane.dangerConfirm("Delete Price Category Detail !")
-                            .confirm(function () {
-                                $scope.model.deletePriceCategoryDetail(items, $index);
-                            })
-                            .discard(function () {
-                                console.log('discard fail');
-                            });
 
-                };
 
                 //---------------------item unit ---------------------
                 //save item units
@@ -216,9 +168,16 @@
                             && $scope.model.itemUnitData.qty
                             && $scope.model.itemUnitData.salePriceNormal
                             && $scope.model.itemUnitData.salePriceRegister) {
-                        $scope.model.saveItemUnit();
-                        $scope.itemType = null;
-                        $scope.itemUnit = null;
+                        var check = true;
+                        if (!$scope.model.userPermission.add) {
+                            ch = false;
+                            Notification.error('you have no permission to save new Item unit !');
+                        }
+                        if (check) {
+                            $scope.model.saveItemUnit();
+                            $scope.itemType = null;
+                            $scope.itemUnit = null;
+                        }
                     }
 
                 };
@@ -238,7 +197,14 @@
 
                 //edit item units
                 $scope.ui.editeItemUnits = function (itemsUnits, $index) {
-                    $scope.model.editeItemUnits(itemsUnits, $index);
+                    var check = true;
+                    if (!$scope.model.userPermission.update) {
+                        ch = false;
+                        Notification.error('you have no permission to edit Item unit !');
+                    }
+                    if (check) {
+                        $scope.model.editeItemUnits(itemsUnits, $index);
+                    }
                 };
 
                 $scope.ui.getItemType = function (model) {
@@ -246,24 +212,37 @@
                     $scope.itemUnit = $scope.model.item(model).unit;
                     $scope.itemObject = $scope.model.item(model);
                     //item selected get item wise item units
-                    console.log('controller');
                     $scope.model.loadItemUnitByItem(model);
                 };
 
                 //delete item units
                 $scope.ui.deleteItemUnits = function (itemsUnits, $index) {
-                    $scope.model.deleteItemUnits(itemsUnits, $index);
+                    var check = true;
+                    if (!$scope.model.userPermission.delete) {
+                        ch = false;
+                        Notification.error('you have no permission to delete item unit !');
+                    }
+                    if (check) {
+                        $scope.model.deleteItemUnits(itemsUnits, $index);
+                    }
                 };
 
                 //--------------------- package ---------------------
                 //save package items
                 $scope.ui.addPackageItem = function () {
-                    if (!$scope.model.packageData.packages) {
-                        Notification.error("select package item");
-                    } else if (!$scope.model.packageData.item) {
-                        Notification.error("select item");
-                    } else if ($scope.model.packageData.packages && $scope.model.packageData.item) {
-                        $scope.model.addPackageItem();
+                    var check = true;
+                    if (!$scope.model.userPermission.add) {
+                        ch = false;
+                        Notification.error('you have no permission to save new package Item !');
+                    }
+                    if (check) {
+                        if (!$scope.model.packageData.packages) {
+                            Notification.error("select package item");
+                        } else if (!$scope.model.packageData.item) {
+                            Notification.error("select item");
+                        } else if ($scope.model.packageData.packages && $scope.model.packageData.item) {
+                            $scope.model.addPackageItem();
+                        }
                     }
                 };
 
@@ -274,36 +253,28 @@
 
                 //consumable Item
                 $scope.ui.deleteConsumableItem = function (index) {
-                    ConfirmPane.dangerConfirm("Delete Selected Consumable Item !")
-                            .confirm(function () {
-                                $scope.model.deleteConsumableItem(index)
-                                        .then(function () {
-                                            Notification.success("Consumable Item Delete Success");
-                                        }, function () {
-                                            Notification.error("Consumable Item Delete Fail");
-                                        });
+                    var check = true;
+                    if (!$scope.model.userPermission.delete) {
+                        check = false;
+                        Notification.error('you have no permission to delete Consumable Item !');
+                    }
+                    if (check) {
+                        ConfirmPane.dangerConfirm("Delete Selected Consumable Item !")
+                                .confirm(function () {
+                                    $scope.model.deleteConsumableItem(index)
+                                            .then(function () {
+                                                Notification.success("Consumable Item Delete Success");
+                                            }, function () {
+                                                Notification.error("Consumable Item Delete Fail");
+                                            });
 
-                            })
-                            .discard(function () {
-                                console.log('discard fail');
-                            });
+                                })
+                                .discard(function () {
+                                    console.log('discard fail');
+                                });
+                    }
                 };
-                //deleteItemCheckDetail
-                $scope.ui.deleteItemCheckDetail = function (index) {
-                    ConfirmPane.dangerConfirm("Delete Selected Item Chech Detail  !")
-                            .confirm(function () {
-                                $scope.model.deleteItemCheckDetail(index)
-                                        .then(function () {
-                                            Notification.success("Item Check Detail Delete Success");
-                                        }, function () {
-                                            Notification.error("Item Check Detail Delete Fail");
-                                        });
-
-                            })
-                            .discard(function () {
-                                console.log('discard fail');
-                            });
-                };
+               
                 $scope.ui.selectPackageItem = function (model) {
                     $scope.model.selectPackageItem(model);
                 };
